feat(ui): add admin layout sidebar and writer dashboard nav
